Move error_caused_by to common.py in tests
