+++ qcodes/tests/common.py
@@ -1,7 +1,11 @@
-from typing import Callable, Type
+from typing import Callable, Type, TYPE_CHECKING
 from functools import wraps
 from time import sleep
 import cProfile
+
+
+if TYPE_CHECKING:
+    from _pytest._code.code import ExceptionInfo
 
 
 def strip_qc(d, keys=('instrument', '__class__')):
@@ -89,3 +93,19 @@ def profile(func):
         profiler.dump_stats(profile_filename)
         return result
     return wrapper
+
+
+def error_caused_by(excinfo: 'ExceptionInfo', cause: str) -> bool:
+    """
+    Helper function to figure out whether an exception was caused by another
+    exception with the message provided.
+
+    Args:
+        excinfo: the output of with pytest.raises() as excinfo
+        cause: the error message or a substring of it
+    """
+    chain = excinfo.getrepr().chain
+    cause_found = False
+    for link in chain:
+        cause_found = cause_found or cause in str(link[1])
+    return cause_found
